lesson 21 (django)/pyproject.toml
[project]
name = "django-test-project"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "crispy-bootstrap5>=2026.3",
    "django>=6.0.4",
    "django-crispy-forms>=2.6",
    "django-debug-toolbar>=6.3.0",
    "pillow>=12.2.0",
    "psycopg2>=2.9.12",
]
